+++ CHANGELOG.md
@@ -31,6 +31,7 @@ Changelog](https://keepachangelog.com/en/1.1.0).
     policy and validator scripts.
   * Added CLI endpoint `init-candidate-permission-token` to initialise the
     candidate permission token mechanism.
+  * Added arbitrary onchain data storage mechnism.
 
 ## Changed
   * `init` now prints more human friendly message, when selected genesis utxo